+++ selfdrive/ui/qt/onroad.h
@@ -81,6 +81,9 @@ class AnnotatedCameraWidget : public CameraWidget {
   Q_PROPERTY(float enginerpm MEMBER enginerpm);
   Q_PROPERTY(bool buttonColorSpeed MEMBER buttonColorSpeed);
   Q_PROPERTY(float distanceTraveled MEMBER distanceTraveled);
+  Q_PROPERTY(bool left_blinker MEMBER left_blinker);
+  Q_PROPERTY(bool right_blinker MEMBER right_blinker);
+  Q_PROPERTY(int blinker_rate MEMBER blinker_rate);
 
 public:
   explicit AnnotatedCameraWidget(VisionStreamType type, QWidget* parent = 0);
@@ -116,7 +119,10 @@ private:
   // AleSato stuff
   float enginerpm; 
   bool buttonColorSpeed = false;
-  float distanceTraveled;   
+  float distanceTraveled;
+  bool left_blinker = false;
+  bool right_blinker = false;
+  int blinker_rate = 120; 
 
 protected:
   void paintGL() override;
@@ -130,6 +136,7 @@ protected:
   inline QColor redColor(int alpha = 255) { return QColor(201, 34, 49, alpha); }
   inline QColor whiteColor(int alpha = 255) { return QColor(255, 255, 255, alpha); }
   inline QColor blackColor(int alpha = 255) { return QColor(0, 0, 0, alpha); }
+  inline QColor yellowColor(int alpha = 255) { return QColor(218, 202, 37, alpha); }
 
   double prev_draw_t = 0;
   FirstOrderFilter fps_filter;